+++ src/main/python/programmingtheiot/cda/system/SystemNetOutUtilTask.py
@@ -27,7 +27,7 @@ class SystemNetOutUtilTask(BaseSystemUtilTask):
         )
         self.last = psutil.net_io_counters().bytes_recv
 
-    def getTelemetryValue(self) -> float:
+    def getTelemetryValue(self) -> int:
         """
         Get the telemetry value
